feat(FN-1088): refactor utilisation-report-upload controller to only use one template
