--- e2e-tests/portal/cypress/e2e/pages/utilisation-report-service/uploadPreviousUtilisationReports.js
@@ -1,9 +1,0 @@
-const page = {
-  utilisationReportFileInput: () => cy.get('[data-cy="utilisation-report-file-upload"]'),
-  continueButton: () => cy.get('[data-cy="continue-button"]'),
-  warningText: () => cy.get('[data-cy="warning-text"]'),
-  dueReportsList: () => cy.get('[data-cy="due-reports-list"]'),
-  currentUrl: () => cy.url(),
-};
-
-module.exports = page;
