--- portal/templates/utilisation-report-service/utilisation-report-upload/upload-previous-utilisation-reports.njk
@@ -1,89 +1,0 @@
-{% extends "index.njk" %}
-{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
-{% set nextDueReportPeriod = [dueReports[0].month, " ", dueReports[0].year] | join %}
-{% set genericWarningText = "There are overdue reports, please send them as soon as possible." %}
-{% set warningText %}
-    {{ genericWarningText if dueReports.length > 2 else [nextDueReportPeriod, " report is overdue, please send it as soon as possible."] | join }}
-{% endset %}
-{% block pageTitle %}Utilisation Report Upload{% endblock %}
-{% block content %}
-    {% if errorSummary %}
-        {{ govukErrorSummary({
-          titleText: "There is a problem",
-          errorList: errorSummary,
-          attributes: {
-            'data-cy': 'error-summary'
-          }
-        }) }}
-    {% endif %}
-    <h1 class="govuk-heading-l" data-cy="main-heading">
-        General Export Facility reporting
-    </h1>
-    {{ govukWarningText({
-        text: warningText,
-        iconFallbackText: "Warning.",
-        attributes: {
-        'data-cy': 'warning-text' }
-    }) }}
-    <ul class="govuk-list govuk-list--bullet" data-cy="due-reports-list">
-        {% for dueReport in dueReports.slice(0, -1) %}
-            <li data-cy="list-item-{{ dueReport.month }}-{{ dueReport.year }}__overdue">
-                {{ dueReport.month }} {{ dueReport.year }} report is overdue
-            </li>
-        {% endfor %}
-        <li data-cy="list-item-{{ dueReports.at(-1).month }}-{{ dueReports.at(-1).year }}__now-due">
-            {{ reportPeriod }} report is now due
-        </li>
-    </ul>
-    <div class="govuk-inset-text" data-cy="inset-text">
-        {% if dueReports.length > 2 %}
-            Once you've sent the {{ nextDueReportPeriod }} report, you can send subsequent reports.
-        {% else %}
-            Once you've sent the {{ nextDueReportPeriod }} report, you can send the {{ reportPeriod }} report.
-        {% endif %}
-    </div>
-    <p class="govuk-body">For each issued facility you must provide:</p>
-    <ul class="govuk-list govuk-list--bullet">
-        <li>facility utilisation accurate for the report month-end date</li>
-        <li>the total fees accrued</li>
-        <li>the fee payable to UKEF</li>
-    </ul>
-    <p class="govuk-body">If there has been no movement against a facility you must still report this.</p>
-    <form method="POST" data-cy="form" enctype="multipart/form-data">
-        <input type="hidden" name="_csrf" value="{{ csrfToken }}"/>
-        <div class="govuk-form-group {% if validationError %}govuk-form-group--error{% endif %}">
-            <label class="govuk-label govuk-label--m" for="utilisation-report-file-upload" data-cy="upload-report-text">
-                Upload {{ nextDueReportPeriod }} report
-            </label>
-            <div id="utilisation-report-file-upload-hint" class="govuk-hint">
-                The file must be an XLSX or CSV
-            </div>
-            {% if validationError %}
-                <span class="govuk-error-message" data-cy="utilisation-report-file-upload-error-message">
-                    <span class="govuk-visually-hidden">Error:</span>
-                    {{ validationError.text }}
-                </span>
-            {% endif %}
-            <input
-                class="govuk-file-upload govuk-!-width-one-half"
-                id="utilisation-report-file-upload"
-                name="utilisation-report-file-upload"
-                data-cy="utilisation-report-file-upload"
-                data-tag="govuk-file-upload"
-                type="file"
-                accept=".xlsx, .csv"
-                />
-        </div>
-        <div class="govuk-button-group">
-            <input
-                formaction="?_csrf={{ csrfToken }}"
-                type="submit"
-                class="govuk-button govuk-!-margin-right-1"
-                data-module="govuk-button"
-                value="Continue"
-                data-cy="continue-button"/>
-        </div>
-    </form>
-{% endblock %}
