+++ portal/templates/utilisation-report-service/utilisation-report-upload/utilisation-report-upload.njk
@@ -1,6 +1,12 @@
 {% extends "index.njk" %}
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
+{% set nextDueReport = dueReportsWithDetails[0] %}
+{% set genericWarningText = "There are overdue reports, please send them as soon as possible." %}
+{% set warningText %}
+    {{ genericWarningText if dueReportsWithDetails.length > 2 else [nextDueReport.reportPeriod, " report is overdue, please send it as soon as possible."] | join }}
+{% endset %}
 {% block pageTitle %}Utilisation Report Upload{% endblock %}
 {% block content %}
     {% if errorSummary %}
@@ -15,12 +21,39 @@
     <h1 class="govuk-heading-l" data-cy="main-heading">
         General Export Facility reporting
     </h1>
-    <h2 class="govuk-heading-m" data-cy="sub-heading">
-        {{ reportPeriod }} report is due
-    </h2>
-    <div class="govuk-inset-text">
-        All issued facilities must be updated and sent to UKEF by {{ reportDueDate }}.
-    </div>
+    {# TODO FN-1089 add behaviour for reports being up to date #}
+    {% if dueReportsWithDetails.length === 1 %}
+        <h2 class="govuk-heading-m" data-cy="sub-heading">
+            {{ nextDueReport.reportPeriod }} report is due
+        </h2>
+        <div class="govuk-inset-text" data-cy="report-due-date">
+            All issued facilities must be updated and sent to UKEF by {{ nextDueReportDueDate }}.
+        </div>
+    {% elif dueReportsWithDetails.length > 1 %}
+        {{ govukWarningText({
+        text: warningText,
+        iconFallbackText: "Warning.",
+        attributes: {
+            'data-cy': 'warning-text' }
+        }) }}
+        <ul class="govuk-list govuk-list--bullet" data-cy="due-reports-list">
+            {% for dueReport in dueReportsWithDetails.slice(0, -1) %}
+                <li data-cy="list-item-{{ dueReport.month }}-{{ dueReport.year }}__overdue">
+                    {{ dueReport.reportPeriod }} report is overdue
+                </li>
+            {% endfor %}
+            <li data-cy="list-item-{{ dueReportsWithDetails.at(-1).month }}-{{ dueReportsWithDetails.at(-1).year }}__now-due">
+                {{ dueReportsWithDetails.at(-1).reportPeriod }} report is now due
+            </li>
+        </ul>
+        <div class="govuk-inset-text" data-cy="inset-text">
+            {% if dueReportsWithDetails.length > 2 %}
+                Once you've sent the {{ nextDueReport.reportPeriod }} report, you can send subsequent reports.
+            {% else %}
+                Once you've sent the {{ nextDueReport.reportPeriod }} report, you can send the {{ dueReportsWithDetails.at(-1).reportPeriod }} report.
+            {% endif %}
+        </div>
+    {% endif %}
     <p class="govuk-body">For each issued facility you must provide:</p>
     <ul class="govuk-list govuk-list--bullet">
         <li>facility utilisation accurate for the report month-end date</li>
@@ -31,8 +64,8 @@
     <form method="POST" data-cy="form" enctype="multipart/form-data">
         <input type="hidden" name="_csrf" value="{{ csrfToken }}"/>
         <div class="govuk-form-group {% if validationError %}govuk-form-group--error{% endif %}">
-            <label class="govuk-label govuk-label--m" for="utilisation-report-file-upload">
-                Upload {{ reportPeriod }} report
+            <label class="govuk-label govuk-label--m" for="utilisation-report-file-upload" data-cy="upload-report-text">
+                Upload {{ nextDueReport.reportPeriod }} report
             </label>
             <div id="utilisation-report-file-upload-hint" class="govuk-hint">
                 The file must be an XLSX or CSV
